change int on size_t in 'for'

+++ zelenin.dmitrii/I2/ismaxdecreasingfragment.cpp
@@ -4,7 +4,7 @@ size_t isMaxDecreasingFragment(int* array, size_t count_elements)
 {
   size_t max_decreasing_fragment = 0;
   size_t decreasing_fragment_counter = 0;
-  for (int i = 0; i + 1 < count_elements; i++)
+  for (size_t i = 0; i + 1 < count_elements; i++)
   {
     if (array[i + 1] < array[i])
     {
